+++ catalog/templates/catalog/base.html
@@ -14,7 +14,7 @@
 </head>
 <body>
 <nav class="navbar navbar-expand-lg  navbar-dark bg-dark sticky-top">
-    <a class="navbar-brand ml-3 " href="../">Ask_margot</a>
+    <a class="navbar-brand ml-3 " href={% url 'ask_margot' %}>Ask_margot</a>
     <div class="my_bar collapse navbar-collapse" id="navbarSupportedContent">
         <div class="col ">
             <form>
