+++ catalog/urls.py
@@ -11,9 +11,9 @@ urlpatterns = [
     path('ask/', views.add_new_question, name='ask'),
     path('question/<int:question_id>', views.question_page, name='question'),
     path('settings/', views.settings_view, name='settings'),
-    path('login/', views.login_view, name='login'),
-    path('logout/', views.logout_view, name='logout'),
-    path('registration/', views.registration_view, name='registration'),
+    # path('login/', views.login_view, name='login'),
+    # path('logout/', views.logout_view, name='logout'),
+    # path('registration/', views.registration_view, name='registration'),
 
 
     url(r'^question/(?P<pk>\d+)/like/$',
